+++ tests/template.rs
@@ -5,10 +5,11 @@ const NAME: &str = "xxxxx";
 #[test]
 fn test_0() {
     let stdin = r#"
+
     "#;
     assert_cli::Assert::cargo_binary(NAME)
         .stdin(stdin)
         .stdout()
-//        .is("3")
+        .is("")
         .unwrap();
 }